updated cluster service interface and related types for consistency
