+++ internal/chunk_replicator/default_replicator/default_posix_chunk_replicator.go
@@ -12,13 +12,13 @@ import (
 )
 
 type DefaultChunkReplicator struct {
-	clusterService cluster_service.NewClusterService
+	clusterService cluster_service.ClusterService
 	comm           communication.Communicator
 	ls             log_service.LogService
 }
 
 func NewDefaultChunkReplicator(
-	clusterService cluster_service.NewClusterService,
+	clusterService cluster_service.ClusterService,
 	comm communication.Communicator,
 	ls log_service.LogService,
 ) *DefaultChunkReplicator {
@@ -40,7 +40,7 @@ func (r *DefaultChunkReplicator) ReplicateChunk(chunkID string, data []byte, fac
 		return crepinternal.ErrInsufficientNodes
 	}
 
-	var candidateNodes []cluster_service.SafeNode
+	var candidateNodes []cluster_service.Node
 	selfAddr := r.comm.Address()
 	for _, n := range nodes {
 		if n.Address != selfAddr {
